# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /screen:
    post:
      operationId: createScreen
      tags: [screener]
      summary: Compare two domains with a first-pass traffic screen
      description: Calls Similarweb for recent estimates and Crusoe for an evidence-constrained brief. Each submission consumes provider credits.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScreenInput"
      responses:
        "200":
          description: Completed live screen
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScreenReport"
        "400":
          description: Invalid domains
        "502":
          description: Provider request failed or access is insufficient
...
  /screens/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string }
    get:
      operationId: getScreen
      tags: [screener]
      summary: Retrieve one of my saved screens
      responses:
        "200":
          description: Saved screen
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SavedScreen" }
    delete:
      operationId: deleteScreen
      tags: [screener]
      summary: Delete one of my saved screens
      responses:
        "204": { description: Deleted }
...
components:
  schemas:
...
    ScreenInput:
      type: object
      additionalProperties: false
      required: [targetDomain, requestId]
      properties:
        targetDomain: { type: string, minLength: 4, maxLength: 253 }
        comparisonDomain: { type: string, minLength: 4, maxLength: 253 }
        comparisonDomains:
          type: array
          items: { type: string, minLength: 4, maxLength: 253 }
        requestId: { type: string, minLength: 8, maxLength: 100 }
    TimePoint:
      type: object
      required: [month, visits]
      properties:
        month: { type: string }
        visits: { type: number }
    ShareItem:
      type: object
      required: [label, share]
      properties:
        label: { type: string }
        share: { type: number }
    DomainProfile:
      type: object
      required: [domain, monthlyVisits, averageVisits, trafficChangePercent, channels, geography, topCountry, topChannel]
      properties:
        domain: { type: string }
        monthlyVisits:
          type: array
          items: { $ref: "#/components/schemas/TimePoint" }
        averageVisits: { type: number }
        trafficChangePercent: { type: ["number", "null"] }
        channels:
          type: array
          items: { $ref: "#/components/schemas/ShareItem" }
        geography:
          type: array
          items: { $ref: "#/components/schemas/ShareItem" }
        topCountry:
          type: ["object", "null"]
          properties:
            label: { type: string }
            share: { type: number }
        topChannel:
          type: ["object", "null"]
          properties:
            label: { type: string }
            share: { type: number }
    Finding:
      type: object
      required: [id, severity, title, metric, value, period, source, sourceUrl, whyItMatters, verificationQuestion]
      properties:
        id: { type: string }
        severity: { type: string, enum: [watch, context] }
        title: { type: string }
        metric: { type: string }
        value: { type: string }
        period: { type: string }
        source: { type: string }
        sourceUrl: { type: string }
        whyItMatters: { type: string }
        verificationQuestion: { type: string }
    ScreenReport:
      type: object
      required: [generatedAt, period, sourceNotice, target, comparison, findings, summary, sellerQuestions, limitations]
      properties:
        generatedAt: { type: string }
        id: { type: string }
        period: { type: string }
        sourceNotice: { type: string }
        target: { $ref: "#/components/schemas/DomainProfile" }
        comparison: { $ref: "#/components/schemas/DomainProfile" }
        additionalProfiles:
          type: array
          items: { $ref: "#/components/schemas/DomainProfile" }
        findings:
          type: array
          items: { $ref: "#/components/schemas/Finding" }
        summary: { type: string }
        sellerQuestions:
          type: array
          items: { type: string }
        limitations:
          type: array
          items: { type: string }
    SavedScreen:
      type: object
      required: [id, createdAt, targetDomain, comparisonDomains, report]
      properties:
        id: { type: string }
        createdAt: { type: string }
        targetDomain: { type: string }
        comparisonDomains:
          type: array
          items: { type: string }
        report: { $ref: "#/components/schemas/ScreenReport" }
...