# lib/api-zod/src/generated/types/screenInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ScreenInput {
  /**
     * @minLength 4
     * @maxLength 253
     */
  targetDomain: string;
  /**
     * @minLength 4
     * @maxLength 253
     */
  comparisonDomain?: string;
  /**
     * @items.minLength 4
     * @items.maxLength 253
     */
  comparisonDomains?: string[];
  /**
     * @minLength 8
     * @maxLength 100
     */
  requestId: string;
}